+++ packages/terra-framework-docs/src/terra-dev-site/doc/data-grid/FlowsheetDataGrid.4/Examples.6/FlowsheetWithSections.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { FlowsheetDataGrid } from 'terra-data-grid';
 
 const gridDataJSON = {
@@ -81,11 +81,72 @@ const FlowsheetWithSections = () => {
     const newSections = [...tableSections];
     const sectionIndex = newSections.findIndex(section => section.id === sectionId);
 
+    const sectionToClear = newSections.find(section => section.id === sectionId);
+
+    sectionToClear.rows = sectionToClear.rows.map(row => ({
+      ...row,
+      cells: row.cells.map(cell => ({
+        ...cell,
+        isSelected: false,
+      })),
+    }));
+
     if (sectionIndex > -1) {
       newSections[sectionIndex].isCollapsed = !newSections[sectionIndex].isCollapsed;
     }
 
     setTableSections(newSections);
+  };
+
+  const getClearedSections = useCallback(() => tableSections.map(section => ({
+    ...section,
+    rows: section.rows.map(row => ({
+      ...row,
+      cells: row.cells.map(cell => ({
+        ...cell,
+        isSelected: false,
+      })),
+    })),
+  })), [tableSections]);
+
+  const onCellSelect = useCallback((selectedCell) => {
+    let selectedSection = tableSections.find(section => section.id === selectedCell.sectionId);
+
+    const columnIndex = gridDataJSON.cols.findIndex(col => col.id === selectedCell.columnId);
+    const rowIndex = selectedSection.rows.findIndex(row => row.id === selectedCell.rowId);
+    const previousCell = selectedSection.rows[rowIndex].cells[columnIndex];
+
+    const newSections = getClearedSections();
+
+    //   // If the current cell is the only selected cell, toggle it to unselected. Otherwise, set it to selected.
+    selectedSection = newSections.find(section => section.id === selectedCell.sectionId);
+    selectedSection.rows[rowIndex].cells[columnIndex].isSelected = !previousCell.isSelected;
+    setTableSections(newSections);
+  }, [tableSections, getClearedSections]);
+
+  const handleCellRangeSelection = useCallback((cells) => {
+    const columnIndexesToUpdate = new Set(cells
+      .map(cell => cell.columnId)
+      .map(id => gridDataJSON.cols.findIndex(col => col.id === id)));
+
+    const rowsToUpdate = new Set(cells.map(cell => cell.rowId));
+
+    const newSections = getClearedSections();
+    const selectedSection = newSections.find(section => section.id === cells[0].sectionId);
+
+    selectedSection.rows = selectedSection.rows.map(row => ({
+      ...row,
+      cells: row.cells.map((cell, cellIndex) => ({
+        ...cell,
+        isSelected: columnIndexesToUpdate.has(cellIndex) && rowsToUpdate.has(row.id),
+      })),
+    }));
+
+    setTableSections(newSections);
+  }, [getClearedSections]);
+
+  const onClearSelectedCells = () => {
+    setTableSections(getClearedSections());
   };
 
   return (
@@ -94,6 +155,9 @@ const FlowsheetWithSections = () => {
       columns={gridDataJSON.cols}
       sections={tableSections}
       onSectionSelect={handleSectionSelect}
+      onCellRangeSelect={handleCellRangeSelection}
+      onCellSelect={onCellSelect}
+      onClearSelectedCells={onClearSelectedCells}
     />
   );
 };