+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.0.0-alpha.3 - (December 13, 2023)
+
 * Added
   * Added keyboard navigation.